% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb.R
\name{count_ssb_tags}
\alias{count_ssb_tags}
\title{Count SSB metadata tags}
\usage{
count_ssb_tags(metadata)
}
\value{
data.frame
}
\description{
Count the number of times a tag has been used.
}
\examples{
# Get the metadata
meta <- get_ssb_metadata()
# Count tags
tags <- count_ssb_tags(meta)
}
