% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kartverket.R
\name{find_address_from_point.data.frame}
\alias{find_address_from_point.data.frame}
\title{find_address_from_point (data.frame)}
\usage{
\method{find_address_from_point}{data.frame}(x, radius = 50, crs = 4258, ...)
}
\description{
find_address_from_point (data.frame)
}
