% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kartverket.R
\name{find_address_from_point.list}
\alias{find_address_from_point.list}
\title{find_address_from_point (list)}
\usage{
\method{find_address_from_point}{list}(x, radius = 50, crs = 4258, ..., raw_response = FALSE)
}
\arguments{
\item{x}{A vector or list with latitude and longitude coordinates.}

\item{radius}{An integer value with the radius for the query, in whole
meters.}

\item{crs}{An integer or vector with the input and output coordinate
reference system. Default value is ETRS 89 for latitude and longitude data
(EPSG:4258).}

\item{...}{Additional arguments passed to the API call. See details.}

\item{raw_response}{If TRUE a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\description{
find_address_from_point (list)
}
