% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kartverket.R
\name{find_address_from_point}
\alias{find_address_from_point}
\title{Find addresses within a geographical area}
\usage{
find_address_from_point(x, radius = 50, crs = 4258, ..., raw_response = FALSE)
}
\arguments{
\item{x}{A vector or list with latitude and longitude coordinates.}

\item{radius}{An integer value with the radius for the query, in whole
meters.}

\item{crs}{The coordinate reference system used for the coordinates given to
the function. Default value is ETRS 89 for latitude and longitude data
(EPSG:4258).}

\item{...}{Additional arguments passed to the API call. See details.}

\item{raw_response}{If TRUE a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\value{
data.frame or list
}
\description{
Search for Norwegian addresses within a certain radius from a geographical
point.
}
\details{
Advanced users might want to look into the additional possible parameters.
For a list of all accepted query parameters see \code{\link[=create_params_kv_punktsok]{create_params_kv_punktsok()}}
or refer to the
\href{https://ws.geonorge.no/adresser/v1/#/default/get_sok}{API documentation}
for further details (in Norwegian only).

The function returns a data.frame or list by default. If you set
\code{raw_response} to \code{TRUE}, the raw response from the API will be returned
instead. Note that the response will then be returned silently.
}
\examples{
# Using a vector
df <- find_address_from_point(c(lat = 63.42805, lon = 10.39679))

# Using a list
dl <- find_address_from_point(
  list(c(lat = 59.91364, lon = 10.7508),
       c(lat = 63.42805, lon = 10.39679)))

# Using a data.frame
dl <- find_address_from_point(
    data.frame(lat = c(59.91364, 63.42805),
               lon = c(10.7508, 10.39679)))

}
