% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kartverket.R
\name{get_address_info}
\alias{get_address_info}
\title{Get address info}
\usage{
get_address_info(search = NULL, ..., raw_response = FALSE)
}
\arguments{
\item{search}{A Norwegian address. For example: 'munkegata 1 trondheim'.}

\item{...}{Additional arguments to send to the API call. See details.}

\item{raw_response}{If TRUE a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\value{
data.frame or list
}
\description{
Search for Norwegian addresses in
\emph{The Norwegian Mapping Authority's Cadastre and Land Registry}.
}
\details{
Many users will get by with using the regular \code{search} parameter, but more
advanced users might want to look into the wide range of possible parameters.
For a list of all accepted query parameters see \code{\link[=create_params_kartverket]{create_params_kartverket()}}
or refer to the
\href{https://ws.geonorge.no/adresser/v1/#/default/get_sok}{API documentation}
for further details (in Norwegian only).
The function returns a data.frame or list by default. If you set \code{raw_response}
to \code{TRUE}, the raw response from the API will be returned instead. Note that the
response will then be returned silently.
}
\examples{

# A simple query
df <- get_address_info('munkegata 1 trondheim')

# Use the advanced parameters to more specific
df <- get_address_info(address_name = 'Munkegata', address_number = 1,
         mun_name = 'Trondheim')

\dontrun{
# Multiple adresses
res <- get_address_info(c('munkegata 1 trondheim', 'tromsø gate'))

# A more advanced query
res <- get_address_info(c('munkegata 1 trondheim', 'munkegata 1'),
   mun_code = c(NULL, '5001'))

}
}
