% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klass.R
\name{get_counties}
\alias{get_counties}
\title{Get counties}
\usage{
get_counties(
  year = format(Sys.Date(), "\%Y"),
  simplify = TRUE,
  raw_response = FALSE
)
}
\arguments{
\item{year}{The year for which the codes should be valid for}

\item{raw_response}{If TRUE a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\value{
data.frame or list
}
\description{
Get all Norwegian counties and county codes for a given year.
}
\details{
Fetches all Norwegian counties and county codes for a given year. All years
from 1972 until present are supported. The function will default to
the current year. The function will return a data.frame with all counties.
If more than one year is given, a single data.frame will be returned with all
codes and names for the given years, unless \code{simplify} is set to \code{FALSE}, then
a list will be returned instead.
}
\seealso{
\code{get_municipalities()} for municipality codes and names
}
