% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klass.R
\name{get_countries}
\alias{get_countries}
\title{Get countries}
\usage{
get_countries(
  year = format(Sys.Date(), "\%Y"),
  include_notes = FALSE,
  simplify = TRUE,
  raw_response = FALSE
)
}
\arguments{
\item{year}{The year for which the codes should be valid for.}

\item{include_notes}{If notes should be included or not.}

\item{simplify}{If \code{TRUE} (default), a single data.frame is returned. Ignored
if \code{raw_response} is set to \code{TRUE}.}

\item{raw_response}{If \code{TRUE} a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\value{
data.frame or list
}
\description{
Get all country codes (alpha-3) based on the ISO 3166-1 standard, but adjusted
for data from Statistics Norway.
}
\details{
Fetches all countries and country codes (alpha-3) for a given year based on the
ISO 3166-1 standard. The list includes special codes XUK and XXX for unknown
citizenship and stateless persons, respectively. These codes are used in official
statistics from Statistics Norway.

All years from 1974 until present are supported. The function defaults to
the current year.

If notes are enabled with \code{include_notes}, a column \code{note} will be added to the
data.frame. Notes are stated with a code reference that has the following order
(alpha-2, alpha-3, num-3, SSB-3), for example Norway (NO, NOR, 578, 000). See
\href{https://www.ssb.no/klass/klassifikasjoner/552}{Statistics Norway's webpage} for
more information on the code specification.

The function returns a data.frame by default. If you prefer the output
as a list you can set \code{simplify} to \code{FALSE}. This can be useful to keep
programmatically track of failed queries. If you set \code{raw_response}
to \code{TRUE}, the raw response from the API will be returned together with the
parsed response. Note that the response will then be returned silently.
}
\examples{
# Get country for current year
get_countries()
}
