% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brreg.R
\name{get_entity}
\alias{get_entity}
\title{Get entity}
\usage{
get_entity(
  entity,
  type = c("both", "main", "sub"),
  simplify = TRUE,
  raw_response = FALSE
)
}
\arguments{
\item{entity}{A Norwegian organisation number or company name.}

\item{type}{Type of entity to query. Either a main, sub or both (default).}

\item{simplify}{If \code{TRUE} (default), a single data.frame is returned. Ignored
if \code{raw_response} is set to \code{TRUE}.}

\item{raw_response}{If \code{TRUE} a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\value{
data.frame or list
}
\description{
Fetch information about an entity from the
\emph{Central Coordinating Register for Legal Entities} (Enhetsregisteret)
based on the entity's name or organisation number.
}
\details{
By default \code{get_entity()} queries information from both the registry for
\emph{main entities} and the registry for \emph{sub-entities} in Enhetsregisteret.
This will however in some cases cause the function to make multiple API
calls for the same entity number. To disable this behavior you can use the
\code{type} parameter to query only for main \emph{or} sub entities.

The function returns a data.frame by default. If you prefer the output
as a list you can set \code{simplify} to \code{FALSE}. This can be useful to keep
programmatically track of failed queries. If you set \code{raw_response}
to \code{TRUE}, the raw response from the API will be returned together with the
parsed response. Note that the response will then be returned silently.

See the
\href{https://data.brreg.no/enhetsregisteret/api/docs/index.html}{API documentation}
for further details (in Norwegian only).
}
\examples{
# Get entity by number
df <- get_entity(974760843)

# Get entity by name
df <- get_entity('Riksrevisjonen')

# Get multiple entities
df <- get_entity(c(974760843, 971524960))

# Only look for sub-entities
df <- get_entity(999178197, type = 'sub')
}
