% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norges-bank.R
\name{get_exchange_rate}
\alias{get_exchange_rate}
\title{Get exchange rate}
\usage{
get_exchange_rate(
  currency,
  frequency = c("daily", "monthly", "annual"),
  n_obs = 10,
  start = NULL,
  end = NULL,
  raw_response = FALSE
)
}
\arguments{
\item{currency}{The currency to fetch data for.}

\item{frequency}{Frequency of data. Can be daily, monthly or annual. Defaults to
daily.}

\item{n_obs}{The number of observations to fetch. See details.}

\item{start}{The first date to fetch data for.}

\item{end}{The last date to fetch data for.}

\item{raw_response}{If \code{TRUE} a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\value{
data.frame or list
}
\description{
Fetch exchange rates from \emph{Norges Bank}.
}
\details{
Currency must be formatted as a three letter
\href{https://www.iso.org/iso-3166-country-codes.html}{ISO 3166} code. A list
of available country codes can be found at
\href{https://www.norges-bank.no/en/topics/Statistics/exchange_rates/}{Norges Bank}.
You can fetch observations for several currencies by passing a character vector.
If you want to include all currencies, you can set the \code{currency} argument
to \code{ALL}. Note that the number of observations is limited to 20 when
retrieving all currencies.

By default, the 10 latest observations are returned. You can adjust the number
of observations by setting \code{n_obs} to a positive integer. You can also fetch
exchange rates for a specific time period by setting the arguments \code{start}
and \code{end}. Both arguments accept a date in the format \verb{\%Y-\%m-\%d} (\emph{YYYY-MM-DD}).

The function returns a data.frame by default. If you set \code{raw_response} to
\code{TRUE}, the raw response from the API will be returned instead. Note that the
response will be returned silently when \code{raw_response} is set to \code{TRUE}.
}
\examples{
# Get latest daily exchange rates
get_exchange_rate('EUR')
# Get monthly exchange rates for multiple currencies
get_exchange_rate(c('EUR', 'USD'), frequency = 'monthly')
# Get last 5 observations for all currencies
get_exchange_rate('ALL', n_obs = 5)

}
