% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klass.R
\name{get_industrial_codes}
\alias{get_industrial_codes}
\title{Get standard industrial codes (SIC)}
\usage{
get_industrial_codes(
  year = format(Sys.Date(), "\%Y"),
  include_notes = FALSE,
  simplify = TRUE,
  raw_response = FALSE
)
}
\arguments{
\item{year}{The year for which the codes should be valid for.}

\item{include_notes}{If notes should be included or not.}

\item{simplify}{If \code{TRUE} (default), a single data.frame is returned. Ignored
if \code{raw_response} is set to \code{TRUE}.}

\item{raw_response}{If \code{TRUE} a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\value{
data.frame or list
}
\description{
Get all standard industrial codes (SIC) based on the NACE standard.
}
\details{
All years from 1994 until present are supported. The function defaults to the
current year.

If notes are enables with \code{include_notes}, a column \code{note} will be added to the
data.frame. The function returns a data.frame by default. If you prefer the output
as a list you can set \code{simplify} to \code{FALSE}. This can be useful to keep
programmatically track of failed queries. If you set \code{raw_response}
to \code{TRUE}, the raw response from the API will be returned together with the
parsed response. Note that the response will then be returned silently.
}
