% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klass.R
\name{get_municipalities}
\alias{get_municipalities}
\alias{get_counties}
\alias{get_adm_units}
\title{Get municipalities and counties}
\usage{
get_municipalities(
  year = format(Sys.Date(), "\%Y"),
  include_notes = FALSE,
  simplify = TRUE,
  raw_response = FALSE
)

get_counties(
  year = format(Sys.Date(), "\%Y"),
  include_notes = FALSE,
  simplify = TRUE,
  raw_response = FALSE
)

get_adm_units(
  year = format(Sys.Date(), "\%Y"),
  include_notes = FALSE,
  simplify = TRUE
)
}
\arguments{
\item{year}{The year for which the codes should be valid for}

\item{include_notes}{If notes should be included or not}

\item{simplify}{If \code{TRUE} (default), a single data.frame is returned. Ignored
if \code{raw_response} is set to \code{TRUE}.}

\item{raw_response}{If \code{TRUE} a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\value{
data.frame or list
}
\description{
Get all Norwegian municipalities or counties, and their respective codes,
for a given year.
}
\details{
\code{get_municipalities()} supports all years from 1977, while \code{get_counties()}
supports all years from 1972. \code{get_adm_units()} is a wrapper function to get
both municipality and county codes in the same function call. All three
functions default to the current year.

The functions returns a data.frame by default. If you prefer the the output
as a list you can set \code{simplify} to \code{FALSE}. This can be useful to keep
programmatically track of failed queries. If you set \code{raw_response}
to \code{TRUE}, the raw response from the API will be returned together with the
parsed response. Note that the response will then be returned silently.
}
\examples{

# Get municipality codes
get_municipalities()

# Get county codes
get_municipalities()

# Get both
get_adm_units()
}
