% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb.R
\name{get_ssb_table}
\alias{get_ssb_table}
\title{Get SSB table}
\usage{
get_ssb_table(id)
}
\arguments{
\item{id}{The table ID. Corresponds to the ID variable in the metadata.}
}
\value{
data.frame
}
\description{
Fetch ready-made tables from Statistics Norway (SSB).
}
\details{
The function returns a ready-made table produced by Statistics Norway, based on the
provided table ID. A list of all available IDs can be found with \code{\link[=get_ssb_metadata]{get_ssb_metadata()}}.

Note that the metadata is updated daily at 05.00 and 11.30 AM, which can lead
to the API being unavailable for up to five minutes.

See the \href{https://data.ssb.no/api/}{API documentation} for further details.
}
\examples{
# Get table with ID '1104'
df <- get_ssb_table('1104')
}
\seealso{
\code{\link[=get_ssb_metadata]{get_ssb_metadata()}} to view the metadata.
}
