% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noapi-package.R
\docType{package}
\name{noAPI-package}
\alias{noAPI}
\alias{noAPI-package}
\title{noAPI: R Client For Public Norwegian APIs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Retrive data from public Norwegian APIs through easy-to-use functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Riksrevisjonen/noAPI}
  \item \url{https://riksrevisjonen.github.io/noAPI/}
  \item Report bugs at \url{https://github.com/Riksrevisjonen/noAPI/issues}
}

}
\author{
\strong{Maintainer}: Ove Haugland Jakobsen \email{ohj@riksrevisjonen.no}

Authors:
\itemize{
  \item Aleksander Eilertsen \email{ale@riksrevisjonen.no}
}

Other contributors:
\itemize{
  \item Joachim Sandnes \email{jsn@riksrevisjonen.no} [contributor]
  \item  Riksrevisjonen [copyright holder]
}

}
\keyword{internal}
