% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kartverket-utils.R
\name{create_params_kv_punktsok}
\alias{create_params_kv_punktsok}
\title{Query parameters for Find Address}
\usage{
create_params_kv_punktsok(
  lat,
  lon,
  radius,
  crs,
  ascii_compatible = TRUE,
  size = 1000,
  page = 0
)
}
\arguments{
\item{radius}{An integer value with the radius for the query, in whole
meters.}

\item{crs}{An integer or vector with the input and output coordinate
reference system. Default value is ETRS89 for latitude and longitude data
(EPSG:4258).}

\item{ascii_compatible}{If TRUE (default) the returned data is ASCII
compatible.}

\item{size}{An integer value with the size for each page. Defaults to 1000.}

\item{page}{An integer value with the page to query. Defaults to 0 (the first
page).}
}
\description{
Accepted query parameters for Kartverket's endpoint \verb{/punktsok}.
}
\keyword{internal}
