% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kartverket.R
\name{find_address_from_point.data.frame}
\alias{find_address_from_point.data.frame}
\title{find_address_from_point (data.frame)}
\usage{
\method{find_address_from_point}{data.frame}(x, radius = 50, crs = 4258, ...)
}
\arguments{
\item{x}{A vector or list with latitude and longitude coordinates.}

\item{radius}{An integer value with the radius for the query, in whole
meters.}

\item{crs}{The coordinate reference system used for the coordinates given to
the function. Default value is ETRS 89 for latitude and longitude data
(EPSG:4258).}

\item{...}{Additional arguments passed to the API call. See details.}
}
\description{
find_address_from_point (data.frame)
}
