% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kartverket.R
\name{find_address_from_point.list}
\alias{find_address_from_point.list}
\title{find_address_from_point (list)}
\usage{
\method{find_address_from_point}{list}(x, radius = 50, crs = 4258, ..., raw_response = FALSE)
}
\description{
find_address_from_point (list)
}
