% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kartverket.R
\name{find_address_from_point.numeric}
\alias{find_address_from_point.numeric}
\title{find_address_from_point (numeric)}
\usage{
\method{find_address_from_point}{numeric}(
  x,
  radius = 50,
  crs = 4258,
  closest = FALSE,
  ...,
  raw_response = FALSE
)
}
\arguments{
\item{x}{A vector or list with latitude and longitude coordinates.}

\item{radius}{An integer value with the radius for the query, in whole
meters.}

\item{crs}{An integer or vector with the input and output coordinate
reference system. Default value is ETRS89 for latitude and longitude data
(EPSG:4258).}

\item{closest}{If TRUE only the address that is closest to the geographical
point is returned.}

\item{...}{Additional arguments passed to the API call.}

\item{raw_response}{If TRUE a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\description{
find_address_from_point (numeric)
}
