% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kartverket.R
\name{find_address_from_point.numeric}
\alias{find_address_from_point.numeric}
\title{find_address_from_point (numeric)}
\usage{
\method{find_address_from_point}{numeric}(x, radius = 50, crs = 4258, raw_response = FALSE, ...)
}
\description{
find_address_from_point (numeric)
}
