% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kartverket.R
\name{find_address_from_point}
\alias{find_address_from_point}
\title{Find addresses within a geographical area}
\usage{
find_address_from_point(
  x,
  radius = 50,
  crs = 4258,
  closest = FALSE,
  ...,
  raw_response = FALSE
)
}
\arguments{
\item{x}{A vector or list with latitude and longitude coordinates.}

\item{radius}{An integer value with the radius for the query, in whole
meters.}

\item{crs}{An integer or vector with the input and output coordinate
reference system. Default value is ETRS89 for latitude and longitude data
(EPSG:4258).}

\item{closest}{If TRUE only the address that is closest to the geographical
point is returned.}

\item{...}{Additional arguments passed to the API call.}

\item{raw_response}{If TRUE a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\value{
data.frame or list
}
\description{
Search for Norwegian addresses within a certain radius from a geographical
point.
}
\details{
For most use cases ETRS89 (EPSG:4258) and WSG84 (EPSG:4326) can be considered
equal, but users should be aware that there is a slight shift on both axes
between the two coordinate systems. If high precision is needed \code{crs} should
be set to 4326.

Advanced users might want to look into the additional possible
parameters. For a list of all accepted query parameters see
\code{\link[=create_params_kv_punktsok]{create_params_kv_punktsok()}} or refer to the \href{https://ws.geonorge.no/adresser/v1/#/default/get_sok}{API documentation} for
further details (in Norwegian only).

The function returns a data.frame or list by default. If you set
\code{raw_response} to \code{TRUE}, the raw response from the API will be returned
instead. Note that the response will then be returned silently.
}
\examples{
# Using a vector
df <- find_address_from_point(c(lat = 63.42805, lon = 10.39679))

# Using a list
dl <- find_address_from_point(
  list(c(lat = 59.91364, lon = 10.7508),
       c(lat = 63.42805, lon = 10.39679)))

}
