% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klass.R
\name{get_countries}
\alias{get_countries}
\title{Get countries}
\usage{
get_countries(
  year = format(Sys.Date(), "\%Y"),
  include_notes = FALSE,
  simplify = TRUE,
  raw_response = FALSE
)
}
\arguments{
\item{year}{The year for which the codes should be valid for}

\item{include_notes}{If notes should be included or not}

\item{simplify}{If \code{TRUE} (default), a single data.frame is returned.}

\item{raw_response}{If TRUE a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\value{
data.frame or list
}
\description{
Get all country codes (alpha-3) based on the ISO 3166-1 standard, but adjusted
for data from Statistics Norway.
}
\details{
Fetches all countries and country codes (alpha-3) for a given year based on the
ISO 3166-1 standard. The list includes special codes XUK and XXX for unknown
citizenship and stateless persons, respectively. These codes are used in official
statistics from Statistics Norway.

All years from 1974 until present are supported. The function will default to
the current year. The function will return a data.frame with all countries and
all country codes. If more than one year is given, a single data.frame will be
returned with all codes and names for the given years, unless \code{simplify} is set
to \code{FALSE}, then a list of the same length of the input will be returned instead.
If \code{raw_response} is set to \code{TRUE}, a parsed data.frame will be returned along
with each raw response from the API. It is not possible to return a single
data.frame when \code{raw_response} is set to \code{TRUE}.

If notes are enables with \code{include_notes}, a column \code{note} will be added to the
data.frame. Notes are stated with a code reference that has the following order
(alpha-2, alpha-3, num-3, SSB-3), for example Norway (NO, NOR, 578, 000). See
\href{https://www.ssb.no/klass/klassifikasjoner/552}{Statistics Norway's webpage} for
more information on the code specification.
}
