% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klass.R
\name{get_industrial_codes}
\alias{get_industrial_codes}
\title{Get Standard Industrial Classification (SIC)}
\usage{
get_industrial_codes(
  year = format(Sys.Date(), "\%Y"),
  include_notes = FALSE,
  simplify = TRUE,
  raw_response = FALSE
)
}
\arguments{
\item{year}{The year for which the codes should be valid for.}

\item{include_notes}{If notes should be included or not.}

\item{simplify}{If \code{TRUE} (default), a single data.frame is returned. Ignored
if \code{raw_response} is set to \code{TRUE}.}

\item{raw_response}{If \code{TRUE} a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\value{
data.frame or list
}
\description{
Get all Standard Industrial Classification (SIC) based on the NACE standard.
}
\details{
Standard Industrial Classification (SIC) is primarily a statistical standard
and is the basis for coding units according to the most important activities
in Statistics Norway's Business register and in the Central Coordinating
Register for Legal Entities. SIC2007 is based on NACE Rev. 2. See
\href{https://www.ssb.no/klass/klassifikasjoner/6}{Statistics Norway's webpage} for
more information on the code specification.

All years from 1994 until present are supported. The function defaults to the
current year. If notes are enabled with \code{include_notes}, a column \code{note} will
be added to the data.frame.

The function returns a data.frame by default. If you prefer the output
as a list you can set \code{simplify} to \code{FALSE}. This can be useful to keep
programmatically track of failed queries. If you set \code{raw_response}
to \code{TRUE}, the raw response from the API will be returned together with the
parsed response. Note that the response will then be returned silently.
}
\seealso{
\code{\link[=get_entity]{get_entity()}}
}
