% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klass.R
\name{get_municipalities}
\alias{get_municipalities}
\title{Get municipalities}
\usage{
get_municipalities(
  year = format(Sys.Date(), "\%Y"),
  add_county = FALSE,
  simplify = TRUE,
  raw_response = FALSE
)
}
\arguments{
\item{year}{The year for which the codes should be valid for}

\item{add_county}{If set to \code{TRUE}, add county name and number}

\item{raw_response}{If TRUE a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\value{
data.frame or list
}
\description{
Get all Norwegian municipalities and municipality codes for a given year.
}
\details{
Fetches all Norwegian municipalities and municipality codes for a given year.
All years from 1977 until present are supported. The function will default to
the current year. The function will return a data.frame with all municipalities.
If more than one year is given, a single data.frame will be returned with all
codes and names for the given years, unless \code{simplify} is set to \code{FALSE}, then
a list will be returned instead.

You can add county numbers and names for each municipality by setting \code{add_county}
to \code{TRUE}.
}
\seealso{
\code{get_counties()} for county codes and names
}
