% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brreg.R
\name{get_municipalities}
\alias{get_municipalities}
\title{Get municipalities}
\usage{
get_municipalities(raw_response = FALSE)
}
\arguments{
\item{raw_response}{If TRUE a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\value{
data.frame or list
}
\description{
Fetch the names and keys for Norwegian municipalities from the
\emph{Central Coordinating Register for Legal Entities} (Enhetsregisteret).
}
\examples{
# Get municipalities
df <- get_municipalities()

}
