% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brreg.R
\name{get_municipalities}
\alias{get_municipalities}
\title{Get municipalities}
\usage{
get_municipalities(raw_response = FALSE)
}
\arguments{
\item{raw_response}{If TRUE a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\value{
data.frame or list
}
\description{
Fetch the names and keys for Norwegian municipalities from the
\emph{Central Coordinating Register for Legal Entities} (Enhetsregisteret).
}
\details{
The function returns a data.frame by default. If you set \code{raw_response} to
\code{TRUE}, the raw response from the API will be returned instead. Note that the
response will then be returned silently.

See the
\href{https://data.brreg.no/enhetsregisteret/api/docs/index.html}{API documentation}
for further details (in Norwegian only).
}
\examples{
# Get municipalities
df <- get_municipalities()

}
