% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brreg.R
\name{get_roles}
\alias{get_roles}
\title{Get roles}
\usage{
get_roles(entity, simplify = TRUE, raw_response = FALSE)
}
\arguments{
\item{entity}{A Norwegian organisation number.}

\item{simplify}{If \code{TRUE} (default), a single data.frame is returned. Ignored
if \code{raw_response} is set to \code{TRUE}.}

\item{raw_response}{If \code{TRUE} a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\value{
list
}
\description{
Fetch information about entity roles from the
\emph{Central Coordinating Register for Legal Entities} (Enhetsregisteret) based
on the organisation number.
}
\details{
The function returns a data.frame by default. If you prefer the output
as a list you can set \code{simplify} to \code{FALSE}. This can be useful to keep
programmatically track of failed queries. If you set \code{raw_response}
to \code{TRUE}, the raw response from the API will be returned together with the
parsed response. Note that the response will then be returned silently.

See the
\href{https://data.brreg.no/enhetsregisteret/api/docs/index.html}{API documentation}
for further details (in Norwegian only).
}
\examples{
# Get roles for a single entity
res <- get_roles(974760843)

# Get roles for a mulitiple enities
res <- get_roles(c(974760843, 971524960))

}
\seealso{
\code{get_entity()}. It's not possible to get roles for an entity using a
name search. If you do not have an organisation number, you should first do
a call to \code{get_entity()} and then call \code{get_roles()} using the response from
the first call.
}
