% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brreg.R
\name{get_roles}
\alias{get_roles}
\title{Get roles}
\usage{
get_roles(entity, raw_response = FALSE)
}
\arguments{
\item{entity}{A Norwegian organisation number.}

\item{raw_response}{If TRUE a list of class \code{noAPI} is
returned, including the raw \code{httr2_response}.}
}
\value{
list
}
\description{
Fetch information about entity roles from the
\emph{Central Coordinating Register for Legal Entities}
(Enhetsregisteret) based on the organisation number.
}
\details{
See the
\href{https://data.brreg.no/enhetsregisteret/api/docs/index.html}{API documentation}
for further details (in Norwegian only).
}
\examples{
# Get roles for a single entity
res <- get_roles(974760843)

# Get roles for a mulitiple enities
res <- get_roles(c(974760843, 971524960))

}
