% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb.R
\name{get_ssb_metadata}
\alias{get_ssb_metadata}
\title{Get SSB table metadata}
\usage{
get_ssb_metadata(data_tag = NULL)
}
\arguments{
\item{data_tag}{Optional. Data tag to filter on.}
}
\value{
data.frame
}
\description{
Fetch metadata for ready-made tables from Statistics Norway (SSB).
}
\details{
The function returns an overview of all the available ready-made
tables produced by Statistics Norway.

Note that the metadata is updated daily at 05.00 and 11.30 AM, which can lead
to the API being unavailable for up to five minutes.

See the \href{https://data.ssb.no/api/}{API documentation} for further details.
}
\examples{
# Get all metadata
ssb_metadata <- get_ssb_metadata()

# Get a subset containing certain tags
ssb_metadata_arbeidskraft <- get_ssb_metadata(data_tag = 'arbeidskraft')

}
\seealso{
\code{\link[=get_ssb_table]{get_ssb_table()}} to retrieve the actual tables.
}
