% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb.R
\name{get_ssb_table}
\alias{get_ssb_table}
\title{Get SSB table}
\usage{
get_ssb_table(id)
}
\arguments{
\item{id}{The table ID corresponding with the ID variable in the meta data.}
}
\value{
data.frame
}
\description{
Fetch ready-made tables from Statistics Norway (SSB).
}
\details{
The function returns a ready-made table produced by Statistics Norway, based on a
provided table ID. A list of all available IDs can be found by running
the function \code{\link[=get_ssb_metadata]{get_ssb_metadata()}}.

Note that the meta data is updated daily at 05.00 and 11.30 AM, which can lead to all tables
being unavailable for up to five minutes.

See the \href{https://data.ssb.no/api/}{API documentation} for further details (in Norwegian only).
}
\examples{
# Get data frame of all meta data
ssb_metadata <- get_ssb_metadata()

# Get data frame of table with ID '1104'
table_1104 <- get_ssb_table('1104')
}
